Update simple_fsm and epool
- Epool now support big memory block <=12MBye -1
- Add debug support to simple_fsm

+++ sources/gmsi/service/memory/epool/epool.c
@@ -41,7 +41,7 @@ end_def_class(pool_t)
 
 typedef void pool_item_init_event_handler_t(void *pTarget, 
                                             void *pItem, 
-                                            uint_fast16_t hwItemSize);
+                                            uint_fast32_t hwItemSize);
 
 /*============================ GLOBAL VARIABLES ==============================*/
 /*============================ LOCAL VARIABLES ===============================*/
