+++ sources/gmsi/service/memory/epool/epool.h
@@ -132,7 +132,7 @@ END_EXTERN_CLASS(pool_t)
 
 typedef void pool_item_init_event_handler_t(void *pTarget, 
                                             void *pItem, 
-                                            uint_fast16_t hwItemSize);
+                                            uint_fast32_t hwItemSize);
 
 /*============================ GLOBAL VARIABLES ==============================*/
 /*============================ LOCAL VARIABLES ===============================*/
